hybrid-search/requirements.txt
faiss-cpu==1.8.0
elasticsearch==8.13.0
numpy==1.26.2
